API tidy up. working for release 1.1

+++ RPS_API/models/UserModel.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 
 namespace RPS_API.Models
 {
@@ -8,7 +6,7 @@ namespace RPS_API.Models
         public string Username { get; set; }
         public int TurnsPlayed { get; set; }
         public int Wins { get; set; }
-        public double WinRatio { get; set; }
+        public int WinRatio { get; set; }
 
         public User(string _username, int _turnsPlayed, int _wins)
         {
@@ -17,6 +15,5 @@ namespace RPS_API.Models
             this.TurnsPlayed = _turnsPlayed;
             //winratio is calculated in controller
         }
-
     }
 }
